Add check before start auto tasks, can be globally disabled now

+++ src/index.js
@@ -111,6 +111,11 @@ app.get('/run/updateActiveCampaignsProgress', async function(req, res, next) {
 })
 
 app.get('/run/cleanDB', async function(req, res, next) {
+	let clean_db_enabled = await getSetting('enable_clean_db')
+	if (clean_db_enabled !== 'yes') {
+		console.log(`Can't update campaigns progress because it's disabled`)
+		return false
+	}
 	let beforeDate = new Date()
 	beforeDate.setMinutes(beforeDate.getMinutes() - 100)
 	const deletedRows = await deleteLogs({ created_at: { _lt: beforeDate } })
